Remove the moonscript client
Sure it worked and all but due to the lagann changes in a6e388dbd58f97098c68c7f740080a57b06c3b00
it doesn't work anymore. Not to mention the fact that it really was just
a quick hack and some parts of it literally shell out to curl to get
around the luasocket library lacking some advanced functionality such as
being able to reliably post a json object to a remote endpoint (hence
the shelling out to curl).